Remove deprecated logging and analysis code; add MultivariateNoiseNormalizer and FlexPCA transformers for improved MEG decoding pipelines.
