--- <MagicMock name='mock.copy().update().fpath' id='22579336956960'>.bak
@@ -1,2 +1,0 @@
-[bad_channels] Saving results...
-[bad_channels] task=restingstate: 2 total bad channels
