--- <MagicMock name='mock.copy().update().fpath' id='22579336961328'>.bak
@@ -1,1 +1,0 @@
-[bad_channels] Saving results...
