--- <MagicMock name='mock.copy().update().fpath' id='22579336969728'>.bak
@@ -1,15 +1,0 @@
-[apply_hfc] Starting analysis...
-[apply_hfc] Loading data...
-[apply_hfc] Loading data...
-[apply_hfc] Loaded raw data for task=restingstate
-[apply_hfc] Running analysis...
-[apply_hfc] Computing HFC projections (order=1)
-[apply_hfc] Computed 3 HFC projection(s)
-3 projection items deactivated
-Created an SSP operator (subspace dimension = 3)
-3 projection items activated
-SSP projectors applied...
-[apply_hfc] Applied HFC projections to task data
-[apply_hfc] HFC application complete!
-[apply_hfc] Saving results...
-[apply_hfc] Saving results...
