--- <MagicMock name='mock.copy().update().fpath' id='22579372120720'>.bak
@@ -1,1 +1,0 @@
-[apply_hfc] Saving results...
