--- <MagicMock name='mock.copy().update().fpath' id='22579372121392'>.bak
@@ -1,20 +1,0 @@
-[bad_segments] Starting analysis...
-[bad_segments] Loading data...
-[bad_segments] Loading data...
-[bad_segments] Loaded raw data for task=restingstate
-[bad_segments] Running analysis...
-[bad_segments] Stage=legacy | channel_threshold=0.05, segment_len_sec=0.5
-[bad_segments] Processing task=restingstate (is_noise=False)
-[bad_segments] Detecting bad segments (filter: 0.1-30.0 Hz, seg=0.5s, thresh=0.05)
-Filtering raw data in 1 contiguous segment
-Setting up band-pass filter from 0.1 - 30 Hz
-
-IIR filter parameters
----------------------
-Butterworth bandpass zero-phase (two-pass forward and reverse) non-causal filter:
-- Filter order 16 (effective, after forward-backward)
-- Cutoffs at 0.10, 30.00 Hz: -6.02, -6.02 dB
-
-[bad_segments] Added 1 new bad-segment annotations (0 total BAD annotations on raw)
-[bad_segments] Saving results...
-[bad_segments] Saving results...
