--- <MagicMock name='mock.copy().update().fpath' id='22579372123408'>.bak
@@ -1,20 +1,0 @@
-[bad_channels] Starting analysis...
-[bad_channels] Loading data...
-[bad_channels] Loading data...
-[bad_channels] Loaded raw data for task=restingstate
-[bad_channels] Running analysis...
-[bad_channels] Processing task=restingstate
-[bad_channels] Detecting bad channels (filter: 1.0-100.0 Hz)
-Filtering raw data in 1 contiguous segment
-Setting up band-pass filter from 1 - 1e+02 Hz
-
-IIR filter parameters
----------------------
-Butterworth bandpass zero-phase (two-pass forward and reverse) non-causal filter:
-- Filter order 16 (effective, after forward-backward)
-- Cutoffs at 1.00, 100.00 Hz: -6.02, -6.02 dB
-
-[bad_channels] Detected 0 bad channels: []
-[bad_channels] Total unique bad channels: 0
-[bad_channels] Saving results...
-[bad_channels] Saving results...
