--- <MagicMock name='mock.copy().update().fpath' id='22579372127440'>.bak
@@ -1,1 +1,0 @@
-[bad_segments] Saving results...
