--- <MagicMock name='mock.copy().update().fpath' id='22651775033296'>.bak
@@ -1,1 +1,0 @@
-[regress_ref] Saving results...
