--- <MagicMock name='mock.copy().update().fpath' id='22725956903200'>.bak
@@ -1,1 +1,0 @@
-[regress_ref] Saving results...
