--- <MagicMock name='mock.copy().update().fpath' id='22725956907568'>.bak
@@ -1,13 +1,0 @@
-[regress_ref] Starting analysis...
-[regress_ref] Loading data...
-[regress_ref] Loading data...
-[regress_ref] Loaded raw data for task=restingstate
-[regress_ref] Running analysis...
-[regress_ref] Processing task=restingstate (is_noise=False)
-[regress_ref] Regressing out reference channels...
-[regress_ref] Using standard (time-invariant) regression
-No projector specified for this dataset. Please consider the method self.add_proj.
-No projector specified for this dataset. Please consider the method self.add_proj.
-[regress_ref] Reference regression complete!
-[regress_ref] Saving results...
-[regress_ref] Saving results...
