--- <MagicMock name='mock.copy().update().fpath' id='23017901212064'>.bak
@@ -1,1 +1,0 @@
-[regress_ref] Saving results...
