--- <MagicMock name='mock.copy().update().fpath' id='23017901222480'>.bak
@@ -1,1 +1,0 @@
-[regress_ref] Saving results...
