--- <MagicMock name='mock.copy().update().fpath' id='23017948546112'>.bak
@@ -1,1 +1,0 @@
-[bad_channels] Saving results...
