--- <MagicMock name='mock.copy().update().fpath' id='23017948552496'>.bak
@@ -1,1 +1,0 @@
-[bad_segments] Saving results...
