--- <MagicMock name='mock.copy().update().fpath' id='23017982697760'>.bak
@@ -1,1 +1,0 @@
-[apply_hfc] Saving results...
