--- <MagicMock name='mock.copy().update().fpath' id='23017982702128'>.bak
@@ -1,1 +1,0 @@
-[apply_hfc] Saving results...
