--- <MagicMock name='mock.copy().update().fpath' id='23017982704144'>.bak
@@ -1,2 +1,0 @@
-[bad_channels] Saving results...
-[bad_channels] task=restingstate: 2 total bad channels
