+++ src/mne_opm.egg-info/SOURCES.txt
@@ -2,11 +2,11 @@ LICENSE
 README.md
 pyproject.toml
 src/custom/__init__.py
-src/custom/auto_coreg.py
 src/custom/custom_preproc.py
 src/custom/format_bids.py
 src/custom/run_beamformer.py
-src/custom/zca_test.py
+src/custom/run_decoding.py
+src/custom/transformers.py
 src/custom/preprocessing/__init__.py
 src/custom/preprocessing/_base.py
 src/custom/preprocessing/_bids_utils.py
@@ -49,7 +49,9 @@ tests/test_beamformer_extended.py
 tests/test_config.py
 tests/test_custom_preproc.py
 tests/test_custom_preproc_extended.py
+tests/test_decoding.py
 tests/test_docs.py
+tests/test_event_alignment.py
 tests/test_format_bids.py
 tests/test_format_bids_extended.py
 tests/test_io_and_bids_utils.py
